+++ kernel/src/memory/physical.rs
@@ -69,46 +69,3 @@ impl RefCountEntry {
 todo!()
 	}
 }
-
-/// Conceptually the same as a hypothetical `Arc<[Frame]>`
-///
-/// # Invariants
-///
-/// Internal implementation assumes that two `ArcFrames` cannot overlap unless they came from the same original `ArcFrames`
-/// object
-pub struct ArcFrames {
-	base: Frame,
-	len: usize
-}
-
-impl ArcFrames {
-	unsafe fn new(base: Frame, len: usize) -> Self {
-		Self {
-			base, len
-		}
-	}
-
-	fn split_at(self, n: usize) -> (Self, Self) {
-		assert!(n <= self.len);
-
-		let second_base = self.base + n;
-		let lens = (n, self.len - n);
-
-		todo!("Insert new RC node");
-
-		(Self { base: self.base, len: lens.0 }, Self { base: second_base, len: lens.1 })
-	}
-}
-
-impl Clone for ArcFrames {
-	fn clone(&self) -> Self {
-		todo!("Update RC nodes");
-		Self { .. *self }
-	}
-}
-
-impl Drop for ArcFrames {
-	fn drop(&mut self) {
-		todo!()
-	}
-}
